% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVCOMPhysics.R
\name{as_seed}
\alias{as_seed}
\title{Given one or more spatial locations, generate POINT objects to seed for tracking}
\usage{
as_seed(X, x, y, z = 0, time = X$get_time()[1], crs = 4326)
}
\arguments{
\item{X}{FVCOM_Physics object}

\item{x}{sf POINT or numeric one of more x coordinates such as longitude
if sf POINT class then y, z, time crs are ignored.}

\item{y}{numeric one of more y coordinates such as latitude}

\item{z}{numeric one of more z coordinates such as depth}

\item{time}{POSIXct one of more time stamps (UTC)}

\item{crs}{the CRS for the input [x,y] coordinates}
}
\value{
sf POINT object suitable for tracking
}
\description{
Given one or more spatial locations, generate POINT objects to seed for tracking
}
