% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{bathymetry}
\alias{bathymetry}
\title{Retrieve bathymetry "sea_floor_depth_below_geoid" in meters
with positive downward.}
\usage{
bathymetry(x, form = c("table", "raster", "mesh")[1])
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
something
}
\description{
Retrieve bathymetry "sea_floor_depth_below_geoid" in meters
with positive downward.
}
