% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CascoBayPhysics.R
\name{CascoBay_Physics}
\alias{CascoBay_Physics}
\title{A class for navigating  FVCOM datasets}
\description{
R6 base class for Boston Innundation
}
\section{Super class}{
\code{\link[fvcom:FVCOM_Physics]{fvcom::FVCOM_Physics}} -> \code{CascoBay}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CascoBay-new}{\code{CascoBay_Physics$new()}}
\item \href{#method-CascoBay-clone}{\code{CascoBay_Physics$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="append_bounds"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-append_bounds'><code>fvcom::FVCOM_Physics$append_bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="char_size"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-char_size'><code>fvcom::FVCOM_Physics$char_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="get_crs"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-get_crs'><code>fvcom::FVCOM_Physics$get_crs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="get_time"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-get_time'><code>fvcom::FVCOM_Physics$get_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="mesh_depth"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-mesh_depth'><code>fvcom::FVCOM_Physics$mesh_depth()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="mesh_metrics"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-mesh_metrics'><code>fvcom::FVCOM_Physics$mesh_metrics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="plot"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-plot'><code>fvcom::FVCOM_Physics$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="print"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-print'><code>fvcom::FVCOM_Physics$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="random_points"><a href='../../fvcom/html/FVCOM_Physics.html#method-FVCOM_Physics-random_points'><code>fvcom::FVCOM_Physics$random_points()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CascoBay-new"></a>}}
\if{latex}{\out{\hypertarget{method-CascoBay-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascoBay_Physics$new(
  filename,
  origin = as.POSIXct("2018-05-01T00:00:00", tz = "UTC"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{character either a filename or OpenDAP URL}

\item{\code{origin}}{POSIXct timestamp indicating the start of the experiment}

\item{\code{...}}{arguments for FVCOM_physics}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CascoBay-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CascoBay-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascoBay_Physics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
