% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CascoBayPhysics.R
\name{CascoBayPhysics}
\alias{CascoBayPhysics}
\title{Instantiate a FVCOM_Physics R6 object}
\usage{
CascoBayPhysics(
  filename = "/mnt/ecocast/coredata/cascobay/cas3_May2018.nc",
  ...
)
}
\arguments{
\item{filename}{character either a filename or OpenDAP URL}

\item{...}{other arguments passed to \code{FVCOM_Physics$new(filename, ...)}}
}
\value{
FVCOM_Physics R6 Reference Object
}
\description{
Instantiate a FVCOM_Physics R6 object
}
