% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{default_template}
\alias{default_template}
\title{Retrieve a raster template that covers the provided mesh}
\usage{
default_template(x, res = c(0.01, 0.01), crs = sf::st_crs(x)[["epsg"]],
  ...)
}
\arguments{
\item{x}{the mesh geometry}

\item{res}{a one or two element numeric of the cell resolution.  If one element
then it is used in x and y directions.}

\item{crs}{character string of the mesh}

\item{...}{further arguments for \code{\link[raster]{raster}}}
}
\value{
a raster layer
}
\description{
Retrieve a raster template that covers the provided mesh
}
