% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvcom-package.R
\docType{package}
\name{fvcom-package}
\alias{fvcom-package}
\alias{fvcom}
\title{fvcom-package}
\description{
Convenience tools for accessing FVCOM NetCDF files
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/fvcom}
}

}
