% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{fvcom_bounds}
\alias{fvcom_bounds}
\title{Determine if elements on the boundary, and if so if open or closed}
\usage{
fvcom_bounds(x, threshold = 1)
}
\arguments{
\item{x}{ncdf4 object}

\item{threshold}{numeric, the depth at which to consider an edge open (deep) or closed (shallow)}
}
\value{
tibble with four variables \code{elem}, \code{h_center}, \code{edge}, \code{bound}
}
\description{
Determine if elements on the boundary, and if so if open or closed
}
