% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_count}
\alias{fvcom_count}
\title{Retrieve a count of nodes or elements}
\usage{
fvcom_count(x, where = c("nodes", "elems", "siglay", "siglev", "time")[1])
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{where}{character, either 'nodes'( or 'node'), 'elems' (or 'elem'),
'siglay', 'siglev', or 'time'}
}
\value{
number of nodes or elements
}
\description{
Retrieve a count of nodes or elements
}
