% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_crs}
\alias{fvcom_crs}
\title{Retrieve the CRS}
\usage{
fvcom_crs(what = "lonlat")
}
\arguments{
\item{what}{character either 'xy', 'lonlat' or 'longlat'.  Actually, if not 'xy'
then 'lonlat' is returned.}
}
\value{
character CRS
}
\description{
Retrieve the CRS
}
