% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_index}
\alias{fvcom_index}
\title{Retrieve the index closest to the specified value(s) for 1-d dimensions.}
\usage{
fvcom_index(x, vals = 0, what = c("time")[1])
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vals}{numeric or POSIXct, one or more values to convert to indices}

\item{what}{character, one of the objects one-dimension names}
}
\value{
vector of indices closest to the specifed val - one for each val
}
\description{
Retrieve the index closest to the specified value(s) for 1-d dimensions.
}
