% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_nav}
\alias{fvcom_nav}
\title{Retrieve the navigation elements, start and count, for extracting slabs of data.}
\usage{
fvcom_nav(
  x,
  vals = c(0, 1),
  what = c("siglay", "time", "siglev")[1],
  indexed = FALSE
)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vals}{two element numeric or POSIXct, start and stop values}

\item{what}{character, one of the objects dimension names}

\item{indexed}{logical, if TRUE then siglay and time are 1-based indices,
but if FALSE then siglay and time are actual values to be located.}
}
\value{
two element vector of start and count
}
\description{
Retrieve the navigation elements, start and count, for extracting slabs of data.
}
