% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_nodes}
\alias{fvcom_nodes}
\title{Retrieve node locations}
\usage{
fvcom_nodes(x, what = c("lonlat", "xy")[1],
  index = seq_len(fvcom_count(x, what = "node")))
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{what}{character, either 'lonlat' (degrees) or 'xy' (meters)}

\item{index}{numeric, a 1-based index of desired nodes}
}
\value{
data frame (tibble) of node locations
}
\description{
Retrieve node locations
}
