% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_random}
\alias{fvcom_random}
\title{Generate random points within the element or node space.}
\usage{
fvcom_random(
  x,
  n = 1,
  what = c("longlat", "xy")[1],
  form = c("tibble", "sf")[1],
  ...
)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{n}{the number of points to generate}

\item{what}{either 'lonlat' or 'xy'}

\item{form}{character either 'tibble' or 'sf'}

\item{...}{arguments for \code{\link{fvcom_time}}}
}
\value{
tibble of
\itemize{
\item{x x coordinate, possibly lon}
\item{y y coordinate, possibly lat}
\item{z z coordinate as sigma value} 
\item{time time ast POSIXct}
}
or an sf POINT object where xy are the geometry
}
\description{
A random point may fall outside of the envelope of elements/nodes since
we use a simple bounding box for each dimension.  Be sure to check your
points, if that is important, with \code{\link[sf]{st_contains}}.
}
