% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_sample}
\alias{fvcom_sample}
\title{Generate a random sample of element [elem, siglay,time] or node [node, siglev, time] indices}
\usage{
fvcom_sample(x, n = 1, where = c("elems", "nodes")[1])
}
\arguments{
\item{x}{ncdf4 FVCOM object}

\item{n}{numeric, the number of IDs to retrieve}

\item{where}{character, either 'nodes' or 'elems' (default)}
}
\value{
named vector of random indices
}
\description{
Generate a random sample of element [elem, siglay,time] or node [node, siglev, time] indices
}
