% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_time}
\alias{fvcom_time}
\title{Get the time dimension values.}
\usage{
fvcom_time(x, origin = as.POSIXct("1858-11-17 00:00:00", tz = "UTC"),
  internal = FALSE)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{origin}{POSIXct start date}

\item{internal}{logical, if TRUE use the internal times, otherwise assume on the hour}
}
\value{
vector of POSIXct times
}
\description{
Note the the models are produced hourly, so we assume on the hour.  To get the
internally recorded hour set internal argument to TRUE.
}
