% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_elem_mesh}
\alias{get_elem_mesh}
\title{Retrieve the node-based mesh with assigned values}
\usage{
get_elem_mesh(x, vars = c("u", "v"), mesh = get_elem_mesh_geometry(x),
  ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vars}{character name of the variable}

\item{mesh}{the mesh geometry}

\item{...}{further arguments for selecting \code{get_node_var}}
}
\value{
sf object with mesh geometry and the var
}
\description{
Retrieve the node-based mesh with assigned values
}
