% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_element_nodes}
\alias{get_element_nodes}
\title{Given an element, get its neighboring nodes}
\usage{
get_element_nodes(x, elem = 1:3)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{elem}{numeric, element ID}
}
\value{
4 column matrix where columns are [elemID, pID1, pID2, nodeID3]
}
\description{
Given an element, get its neighboring nodes
}
