% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_element_vars}
\alias{get_element_vars}
\title{Get one or more variables by [element, siglay, time]}
\usage{
get_element_vars(
  x,
  vars = c("u", "v", "ww"),
  indices = matrix(c(1, 100, 200, 300, 1, 3, 5, 8, 1, 2, 3, 4), ncol = 3)
)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vars}{character one or more variable names to retrieve}

\item{indices}{3 element indices of the element volume to query or a matrix
of three columns}
}
\value{
matrix
}
\description{
Get one or more variables by [element, siglay, time]
}
