% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_mesh}
\alias{get_mesh}
\title{Retrieve node or element based polygon mesh}
\usage{
get_mesh(x, where = c("nodes", "elems")[1], ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{where}{character either 'nodes' or 'elems'}

\item{...}{further arguments for \code{\link{get_node_mesh}} or \code{\link{get_elem_mesh}}}
}
\value{
sf object with mesh geometry and the variables
}
\description{
Retrieve node or element based polygon mesh
}
