% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_mesh_geometry}
\alias{get_mesh_geometry}
\title{Compute mesh (polygons) for nodes or elements}
\usage{
get_mesh_geometry(x, where = c("nodes", "elems"), ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{where}{character, either 'nodes' or "elems"}

\item{...}{further arguments for \code{\link{get_elem_mesh_geometry}} or \code{\link{get_nodes_mesh_geometry}}}
}
\value{
sf collection of POLYGON with the following variable
\itemize{
  \item{p1, p2, p3 the point identifiers for the three points defining the polygon}
  \item{geometry 4 point polygon (closed so first and last the same)}
}
}
\description{
Compute mesh (polygons) for nodes or elements
}
