% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_node_mesh}
\alias{get_node_mesh}
\title{Retrieve the mesh with assigned values}
\usage{
get_node_mesh(x, vars = "zeta", mesh = get_node_mesh_geometry(x),
  fun = mean, ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vars}{character name of the variable}

\item{mesh}{the mesh geometry}

\item{fun}{function to summarize the variable for each polygon}

\item{...}{further arguments for selecting \code{get_node_var}}
}
\value{
sf object with mesh geometry and the var (computed as mean of nodes)
}
\description{
Retrieve the mesh with assigned values
}
