% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_node_var}
\alias{get_node_var}
\title{Get variables for a set of nodes.  This is a convenience
function that provides a means for getting a vector of values at a given
sigma layer (or level) and time.  Note that the dimensions a
variable uses varies - see \code{\link{list_vars}}.}
\usage{
get_node_var(
  x,
  var = "zeta",
  node = seq_len(fvcom_count(x, "nodes")),
  y = 1,
  time = 1,
  y_transform = FALSE
)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{var}{charcater the name of the variable.  Not all variables are handled
by this function.}

\item{node}{integer one or more 1-based node ID(s).  These must form
a contiguous sequence of node identifers}

\item{y}{integer a single 1-based index into siglay, or sigelev}

\item{time}{either a single 1-based integer of POSIXct}

\item{y_transform}{logical, if convert y from sigma to integer?
Note that time is auto-converted}
}
\value{
tibble of node id and values - one row for each node
}
\description{
Get variables for a set of nodes.  This is a convenience
function that provides a means for getting a vector of values at a given
sigma layer (or level) and time.  Note that the dimensions a
variable uses varies - see \code{\link{list_vars}}.
}
