% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_var}
\alias{get_var}
\title{Get variables for nodes or elements.}
\usage{
get_var(x, where = c("nodes", "elems"), ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{where}{character - either 'node' or 'elem'}

\item{...}{further arguments for \code{\link{get_node_var}} or \code{\link{get_elem_var}}}
}
\value{
tibble - see \code{\link{get_node_var}} or \code{\link{get_elem_var}}
}
\description{
Get variables for nodes or elements.
}
