% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_vars}
\alias{get_vars}
\title{Get variables for nodes or elements.  It is not possible to get a mix of
variables - that is some from nodes and some from elements - in one call.}
\usage{
get_vars(x, vars = c("u", "v"), lut = list_vars(x), ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vars}{character - one or more variables}

\item{lut}{a table of variable descriptions - see \code{\link{list_vars}}}

\item{...}{further arguments for \code{\link{get_node_var}} or \code{\link{get_elem_var}}}
}
\value{
list of variables - see \code{\link{get_node_var}} or \code{\link{get_elem_var}}
}
\description{
Get variables for nodes or elements.  It is not possible to get a mix of
variables - that is some from nodes and some from elements - in one call.
}
