% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalPhysics.R
\name{Global_Physics}
\alias{Global_Physics}
\title{A class for navigating  FVCOM datasets}
\description{
R6 base class for the globe
}
\section{Super class}{
\code{\link[fvcom:FVCOM_Physics]{fvcom::FVCOM_Physics}} -> \code{Global_Physics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Global_Physics$new()}}
\item \href{#method-clone}{\code{Global_Physics$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="append_bounds">}\href{../../fvcom/html/FVCOM_Physics.html#method-append_bounds}{\code{fvcom::FVCOM_Physics$append_bounds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="char_size">}\href{../../fvcom/html/FVCOM_Physics.html#method-char_size}{\code{fvcom::FVCOM_Physics$char_size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="get_crs">}\href{../../fvcom/html/FVCOM_Physics.html#method-get_crs}{\code{fvcom::FVCOM_Physics$get_crs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="get_time">}\href{../../fvcom/html/FVCOM_Physics.html#method-get_time}{\code{fvcom::FVCOM_Physics$get_time()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="mesh_metrics">}\href{../../fvcom/html/FVCOM_Physics.html#method-mesh_metrics}{\code{fvcom::FVCOM_Physics$mesh_metrics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="plot">}\href{../../fvcom/html/FVCOM_Physics.html#method-plot}{\code{fvcom::FVCOM_Physics$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="print">}\href{../../fvcom/html/FVCOM_Physics.html#method-print}{\code{fvcom::FVCOM_Physics$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fvcom" data-topic="FVCOM_Physics" data-id="random_points">}\href{../../fvcom/html/FVCOM_Physics.html#method-random_points}{\code{fvcom::FVCOM_Physics$random_points()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Global_Physics$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments for FVCOM_physics}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Global_Physics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
