% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalPhysics.R
\name{GlobalPhysics}
\alias{GlobalPhysics}
\title{Instantiate a FVCOM_Physics R6 object}
\usage{
GlobalPhysics(
  filename = file.path("http://www.smast.umassd.edu:8080",
    "thredds/dodsC/models/fvcom/NECOFS/Forecasts",
    "NECOFS_FVCOM_OCEAN_GLOBAL_FORECAST.nc"),
  ...
)
}
\arguments{
\item{filename}{character either a filename or OpenDAP URL}

\item{...}{other arguments passed to \code{FVCOM_Physics$new(filename, ...)}}
}
\value{
FVCOM_Physics R6 Reference Object
}
\description{
Instantiate a FVCOM_Physics R6 object
}
