% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOMPhysics.R
\name{GOMPhysics}
\alias{GOMPhysics}
\title{Instantiate a FVCOM_Physics R6 object}
\usage{
GOMPhysics(
  filename = file.path("http://www.smast.umassd.edu:8080",
    "thredds/dodsC/models/fvcom/NECOFS/Forecasts", "NECOFS_GOM4_FORECAST.nc"),
  ...
)
}
\arguments{
\item{filename}{character either a filename or OpenDAP URL}

\item{...}{other arguments passed to \code{FVCOM_Physics$new(filename, ...)}}
}
\value{
FVCOM_Physics R6 Reference Object
}
\description{
Instantiate a FVCOM_Physics R6 object
}
