% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{is_ncdf4}
\alias{is_ncdf4}
\title{Test is an object is of ncdf4 class}
\usage{
is_ncdf4(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical TRUE if the object is a valid ncdf4 class
}
\description{
Test is an object is of ncdf4 class
}
