% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{list_vars}
\alias{list_vars}
\title{Generate a listing of variables and dimensions}
\usage{
list_vars(x)
}
\arguments{
\item{x}{FVCOM ncdf4 object}
}
\value{
handy tibble
}
\description{
Generate a listing of variables and dimensions
}
