% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassBayPhysics.R
\name{MassBayPhysics}
\alias{MassBayPhysics}
\title{Instantiate a FVCOM_Physics R6 object}
\usage{
MassBayPhysics(
  filename = file.path("http://www.smast.umassd.edu:8080",
    "thredds/dodsC/models/fvcom/NECOFS/Archive", "NECOFS_MASS_BAY/2020/mbn_202012.nc"),
  ...
)
}
\arguments{
\item{filename}{character either a filename or OpenDAP URL}

\item{...}{other arguments passed to \code{FVCOM_Physics$new(filename, ...)}}
}
\value{
FVCOM_Physics R6 Reference Object
}
\description{
Instantiate a FVCOM_Physics R6 object
}
