% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mgrepl}
\alias{mgrepl}
\title{Perform grepl on multiple patterns; it's like  AND-ing or OR-ing successive grepl statements.}
\usage{
mgrepl(pattern, x, op = `|`, ...)
}
\arguments{
\item{pattern}{character vector of patterns}

\item{x}{the character vector to search}

\item{op}{logical vector operator back quoted, defaults to `|`}

\item{...}{further arguments for \code{grepl} like \code{fixed} etc.}
}
\value{
logical vector
}
\description{
Adapted from https://stat.ethz.ch/pipermail/r-help/2012-June/316441.html
}
