% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{mur_template}
\alias{mur_template}
\title{Retrieve a raster template that covers the provided mesh and aligns with MUR SST pixels}
\usage{
mur_template(x, res = c(0.01, 0.01),
  crs = "+proj=longlat +datum=WGS84 +no_defs", bb = c(-75.69, -56.85,
  35.27, 46.15), ...)
}
\arguments{
\item{x}{the mesh geometry}

\item{res}{a one or two element numeric of the cell resolution.  If one element
then it is used in x and y directions.}

\item{crs}{character string of the mesh}

\item{bb}{numeric, 4 element extent [left, right, bottom, top] by default matching MURSST}

\item{...}{further arguments for \code{\link[raster]{raster}}}
}
\value{
a raster layer
}
\description{
Retrieve a raster template that covers the provided mesh and aligns with MUR SST pixels
}
\seealso{
\href{https://podaac.jpl.nasa.gov/dataset/MUR-JPL-L4-GLOB-v4.1}{MUR SST}
}
