% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{particle_track}
\alias{particle_track}
\title{Track one or more particles}
\usage{
particle_track(
  X,
  P0 = X$random_points(),
  tstep = 60,
  tmax = 3600 * 12,
  reverse = FALSE,
  drag = c(0, 0, 0),
  fixed_z = FALSE,
  clip_z = TRUE,
  show_progress = FALSE,
  verbose = FALSE,
  filename = c("particle_track.gpkg", NA)[2],
  overwrite = TRUE
)
}
\arguments{
\item{X}{FVCOM_Physics object}

\item{P0}{starting point(s), see \code{NE_Physics$random_point}}

\item{tstep}{numeric, seconds between each iteration}

\item{tmax}{numeric, number of seconds to run for}

\item{reverse}{logical, if TRUE run the track reverse in time.  It is an error
to provide the seed point with a timestamp equivalent to the first timestep of the
model.  See \code{\link{random_point}} to set the seed with other timestamps}

\item{drag}{numeric, a three element vector of drag factors in [x, y, z], like a sinking rate in z.
Drag units must be the same as the units of \code{u}, \code{v}, and \code{ww}.}

\item{fixed_z}{logical, if TRUE then depths are fixed to the initial depths, and 
\code{drag} is truncated to two elements.}

\item{clip_z}{logical, if TRUE clip Z positions to be between surface and bathymetric depth.}

\item{show_progress}{logical, if TRUE then show a progress bar}

\item{verbose}{logical, if TRUE output messages (for development)}

\item{filename}{an optional filename to save the track to or NA to skip}

\item{overwrite}{logical, if TRUE allow overwriting of existing files}
}
\value{
sf object of type POINT
}
\description{
Track one or more particles
}
