% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mesh_geometry}
\alias{plot_mesh_geometry}
\title{Given a (small) mesh, plot the mesh and possibly label}
\usage{
plot_mesh_geometry(mesh, ..., label_vertices = TRUE,
  label_polygons = FALSE)
}
\arguments{
\item{mesh}{a mesh produced by \code{\link{get_node_mesh_geometry}}}

\item{...}{further arguments for \code{plot}}

\item{label_vertices}{logical, if TRUE then label the vertices}

\item{label_polygons}{logical, if TRUE label each polygon with it's row number}
}
\description{
Given a (small) mesh, plot the mesh and possibly label
}
