% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVCOMPhysics.R
\name{point_element}
\alias{point_element}
\title{Given a POINT object, determine the elements each point belongs to}
\usage{
point_element(X, p)
}
\arguments{
\item{X}{FVCOM_Physics object}

\item{p}{sf POINT object}
}
\value{
integer index matching the mesh element that contains each point.  For points
  outside of the mesh -1 is returned.
}
\description{
Given a POINT object, determine the elements each point belongs to
}
