% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{rasterize}
\alias{rasterize}
\title{Interpolate a raster from a mesh}
\usage{
rasterize(x, template = default_template(x), ...)
}
\arguments{
\item{x}{the mesh as `sf`` POLYGON}

\item{template}{a raster template see \code{\link{default_template}} and \code{\link{mur_template}}}

\item{...}{further arguments for \code{\link[stars]{st_rasterize}}}
}
\value{
a raster layer
}
\description{
Interpolate a raster from a mesh
}
\examples{
\dontrun{
  mesh <- fvcom::get_mesh(x, vars = c("u", "v"), what = "lonlat")
  uv <- lapply(c("u", "v"), function(f) fvcom::rasterize(mesh[f]))
  
}
}
