% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{sigma_index}
\alias{sigma_index}
\title{Convert a time to a sigma level or sigma layer}
\usage{
sigma_index(x, sigma = seq(-1, 0, by = 0.1), what = c("siglev", "siglay"))
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{sigma}{one or more sigma values}

\item{what}{character either 'siglev' or 'siglay'}
}
\value{
1-based index
}
\description{
Convert a time to a sigma level or sigma layer
}
