% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track_add_z}
\alias{track_add_z}
\title{Extract Z from coordinates and add as a variable for one or more tracks}
\usage{
track_add_z(x)
}
\arguments{
\item{x}{sf POINT object, with a \code{track} column.}
}
\value{
the input sf POINT object with added \code{Z} variable (column)
}
\description{
Extract Z from coordinates and add as a variable for one or more tracks
}
