% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{write_track}
\alias{write_track}
\title{Write a track to file}
\usage{
write_track(x, filename = "track.geojson", overwrite = TRUE, crs = 4326, ...)
}
\arguments{
\item{x}{sf track object}

\item{filename}{character, the filename to write to}

\item{overwrite}{logical, if TRUE overwrite existing file}

\item{crs}{NA or crs-castable object (like 4326). If not NA then
transform to the crs just before writing.  (geojson only accepts 4326-ish)}

\item{...}{other arguments for \code{\link[sf]{write_sf}}}
}
\value{
the input object
}
\description{
Write a track to file
}
