% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{distinct_polygons}
\alias{distinct_polygons}
\title{Compute a table of distinct polygons (triangles) by listing the
nodes (or elements).}
\usage{
distinct_polygons(x, what = c("nodes", "elems")[1])
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{what}{character, either 'nodes' or "elems"}
}
\value{
ordered tibble of p1, p2, and p3
}
\description{
Compute a table of distinct polygons (triangles) by listing the
nodes (or elements).
}
