% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{fvcom_crs}
\alias{fvcom_crs}
\title{Retrieve the CRS}
\usage{
fvcom_crs(
  x,
  what = c("lonlat", "xy")[1],
  form = c("wkt", "epsg", "proj")[ifelse(tolower(what[1]) == "xy", 1, 2)]
)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{what}{character either 'xy', 'lonlat' or 'longlat'.  Actually, if not 'xy'
then 'lonlat' is returned.}

\item{form}{character, one of 'wkt', 'epsg' or 'proj' By default 'wkt' if \code{what} is
'xy' and 'epsg' otherwise.}
}
\value{
character CRS
}
\description{
Retrieve the CRS
}
