% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_elem_mesh}
\alias{get_elem_mesh}
\title{Retrieve the element-based polygon mesh with assigned values}
\usage{
get_elem_mesh(x, vars = c("u", "v"), mesh = get_elem_mesh_geometry(x,
  ordered = FALSE), ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vars}{character one or more variable names}

\item{mesh}{the mesh geometry}

\item{...}{further arguments (ignored)}
}
\value{
sf object with mesh geometry and the variables
}
\description{
Retrieve the element-based polygon mesh with assigned values
}
