% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_elem_mesh_geometry}
\alias{get_elem_mesh_geometry}
\title{Compute the mesh (polygons) for elements}
\usage{
get_elem_mesh_geometry(x, what = "lonlat")
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{what}{character either 'lonlat' or 'xy'}
}
\value{
sf collection of POLYGON with the following variable
\itemize{
\item{p1, p2, p3 the node identifiers for the three points define the polygon}
\item{geometry 4 point polygon (closed so first and last the same)}
}
}
\description{
Compute the mesh (polygons) for elements
}
