% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_mesh}
\alias{get_mesh}
\title{Retrieve node or element based polygon mesh}
\usage{
get_mesh(x, vars = c("zeta", "u", "v"), lut = list_vars(x),
  mesh = get_elem_mesh_geometry(x, ordered = FALSE), ...)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vars}{character one or more variable names to retrieve}

\item{lut}{a table of variable descriptions - see \code{\link{list_vars}}}

\item{mesh}{the mesh geometry}

\item{...}{further arguments for \code{\link{get_node_mesh}} or \code{\link{get_elem_mesh}}}
}
\value{
sf object with mesh geometry and the variables
}
\description{
Retrieve node or element based polygon mesh
}
