% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_node_mesh}
\alias{get_node_mesh}
\title{Retrieve the node-based mesh with assigned values}
\usage{
get_node_mesh(
  x,
  vars = "zeta",
  mesh = get_node_mesh_geometry(x, ordered = FALSE),
  fun = mean,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{FVCOM ncdf4 object}

\item{vars}{character one or more variable names}

\item{mesh}{the mesh geometry}

\item{fun}{function to summarize the variable for each polygon}

\item{na.rm}{logical}

\item{...}{further arguments for selecting \code{get_node_var}}
}
\value{
sf object with mesh geometry and the var (computed as fun of nodes)
}
\description{
Retrieve the node-based mesh with assigned values
}
