% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_track}
\alias{plot_track}
\title{Plot a track or series of tracks}
\usage{
plot_track(
  p,
  X = NULL,
  title = "Particle Track",
  filename = c(NA, "particle_track.png")[1],
  ext = p,
  ...
)
}
\arguments{
\item{p}{sf POINT tibble.  If it has a 'track' variable (column) then each is plotted}

\item{X}{NE_Physics object}

\item{title}{character plot title}

\item{filename}{character or NA, optional output file as PNG}

\item{ext}{object that defines plot extent, by default \code{p} See \code{\link[sf]{plot}}}

\item{...}{other arguments for \code{\link[grDevices]{png}}}
}
\description{
Plot a track or series of tracks
}
