% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_z}
\alias{plot_z}
\title{Plot one or more tracks by depth (Z)}
\usage{
plot_z(
  x,
  title = "Particle Track - Z",
  zvar = c("sigma", "depth", "Z")[3],
  xvar = c("path_distance", "time")[1],
  ylab = c(sigma = "Sigma", depth = "Depth (m)", Z = "Depth (m)")[zvar],
  xlab = c(path_distance = "Path Distance (m)", time = "Time")[xvar],
  filename = c(NA, "track-Z.png")[1],
  ...
)
}
\arguments{
\item{x}{sf POINT object with one or more tracks}

\item{title}{char, title to apply to the plot}

\item{zvar}{char, the name of the zvariable}

\item{xvar}{char, the name of the independent variable ("x")}

\item{ylab}{char, y axis label}

\item{xlab}{char,  x axis label}

\item{filename}{NA or char, for optional output}

\item{...}{other arguments for the graphics device (\code{\link[grDevices]{pdf}} or \code{\link[grDevices]{png}})}
}
\description{
Plot one or more tracks by depth (Z)
}
