% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVCOMPhysics.R
\name{random_point}
\alias{random_point}
\title{Generate  listing of one or more random points}
\usage{
random_point(
  X,
  n = 1,
  select_time = c("first", "last", "random")[1],
  zvar = c("siglev", "siglay")[1]
)
}
\arguments{
\item{X}{FVCOM_Physics object}

\item{n}{numeric the number of random points to generate}

\item{select_time}{character, one of 'first' (default), 'last' or 'random'}

\item{zvar}{character, the name of the varoiable to use for z locations}
}
\value{
sf POINT XYZ object with \code{n} features
}
\description{
Generate  listing of one or more random points
}
