% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{rasterize}
\alias{rasterize}
\title{Interpolate a raster from a mesh}
\usage{
rasterize(x, template = default_template(x), ...)
}
\arguments{
\item{x}{the mesh as `sf`` POLYGON}

\item{template}{a raster template}

\item{...}{further arguments for \code{\link[fasterize]{fasterize}}}
}
\value{
a raster layer
}
\description{
Interpolate a raster from a mesh
}
\examples{
\dontrun{
  mesh <- fvcom::get_mesh(x, vars = c("u", "v"), what = "lonlat")
  uv <- raster::stack(sapply(c("u", "v"), function(f) rasterize(mesh, field = f), simplify = FALSE))
  rasterVis::vectorplot(uv, isField = TRUE, main = 'Surface Currents')
}
}
